Refactor plugin structure
